refactor: restructure into 4-tier architecture(domain, application, ./database, presentation)
